[#82] refactor: 명령을 수행하는 MemberService와 조회를 수행하는 MemberQueryService로 분리
