+++ src/main/java/com/konggogi/veganlife/member/service/MemberService.java
@@ -2,11 +2,7 @@ package com.konggogi.veganlife.member.service;
 
 
 import com.konggogi.veganlife.global.exception.ErrorCode;
-import com.konggogi.veganlife.global.exception.NotFoundEntityException;
-import com.konggogi.veganlife.global.security.exception.MismatchTokenException;
-import com.konggogi.veganlife.global.security.jwt.JwtProvider;
 import com.konggogi.veganlife.global.security.jwt.RefreshToken;
-import com.konggogi.veganlife.global.util.JwtUtils;
 import com.konggogi.veganlife.member.controller.dto.request.MemberInfoRequest;
 import com.konggogi.veganlife.member.domain.Member;
 import com.konggogi.veganlife.member.exception.DuplicateNicknameException;
@@ -18,14 +14,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class MemberService {
     private final MemberRepository memberRepository;
+    private final MemberQueryService memberQueryService;
     private final RefreshTokenRepository refreshTokenRepository;
-    private final JwtUtils jwtUtils;
-    private final JwtProvider jwtProvider;
 
-    @Transactional
     public Member addMember(String email) {
         return memberRepository
                 .findByEmail(email)
@@ -36,25 +30,18 @@ public class MemberService {
                         });
     }
 
-    @Transactional
     public void removeMember(Long memberId) {
-        Member member = findMemberById(memberId);
+        Member member = memberQueryService.findMemberById(memberId);
         memberRepository.delete(member);
     }
 
-    public Member search(Long memberId) {
-        return findMemberById(memberId);
-    }
-
-    @Transactional
     public Member modifyMemberInfo(Long memberId, MemberInfoRequest memberInfoRequest) {
         validateNickname(memberInfoRequest.nickname());
-        Member member = findMemberById(memberId);
+        Member member = memberQueryService.findMemberById(memberId);
         member.updateMemberInfo(memberInfoRequest);
         return member;
     }
 
-    @Transactional
     public void saveRefreshToken(Long memberId, String token) {
         refreshTokenRepository
                 .findRefreshTokenByMemberId(memberId)
@@ -66,24 +53,6 @@ public class MemberService {
                         });
     }
 
-    public String reissueToken(String refreshToken) {
-        jwtUtils.validateToken(refreshToken);
-        Member member = findMemberByToken(refreshToken);
-        return refreshTokenRepository
-                .findRefreshTokenByMemberId(member.getId())
-                .map(
-                        token -> {
-                            if (!token.isSameToken(refreshToken)) {
-                                throw new MismatchTokenException(ErrorCode.MISMATCH_REFRESH_TOKEN);
-                            }
-                            return jwtProvider.createToken(member.getEmail());
-                        })
-                .orElseThrow(
-                        () -> {
-                            throw new NotFoundEntityException(ErrorCode.NOT_FOUND_REFRESH_TOKEN);
-                        });
-    }
-
     private void validateNickname(String nickname) {
         memberRepository
                 .findByNickname(nickname)
@@ -91,17 +60,5 @@ public class MemberService {
                         member -> {
                             throw new DuplicateNicknameException(ErrorCode.DUPLICATED_NICKNAME);
                         });
-    }
-
-    private Member findMemberById(Long memberId) {
-        return memberRepository
-                .findById(memberId)
-                .orElseThrow(() -> new NotFoundEntityException(ErrorCode.NOT_FOUND_MEMBER));
-    }
-
-    private Member findMemberByToken(String token) {
-        return jwtUtils.extractUserEmail(token)
-                .flatMap(memberRepository::findByEmail)
-                .orElseThrow(() -> new NotFoundEntityException(ErrorCode.NOT_FOUND_MEMBER));
     }
 }
